Added HighScore methods to Leaderboard.java
Added saveHighScore and loadHighScore methods
Writes new high score entry to highscores.csv, reads high scores from the CSV as well.  It initializes an empty list to score the highscore objects and reads each line of the file, splits it using a delimiter, and parses the score to an integer (score), string(name), and LocalDate object.  
Also creates a new HighScore object and adds it to the list of high scores.